Move queue-related methods into own modules
